+++ src/main/java/com/github/zastrixarundell/torambot/commands/gameinfo/DyeCommand.java
@@ -5,6 +5,8 @@ import com.github.zastrixarundell.torambot.Values;
 import org.javacord.api.entity.message.embed.EmbedBuilder;
 import org.javacord.api.event.message.MessageCreateEvent;
 import org.javacord.api.listener.message.MessageCreateListener;
+import org.joda.time.DateTime;
+import org.joda.time.Period;
 
 public class DyeCommand implements MessageCreateListener
 {
@@ -26,14 +28,22 @@ public class DyeCommand implements MessageCreateListener
 
         Runnable runnable = () ->
         {
+            DateTime time = new DateTime();
+            Period period = new Period(Values.getLastDyeUpdate(), time);
+
             EmbedBuilder embed = new EmbedBuilder()
                     .setTitle("Latest monthly dyes")
                     .setDescription("Here is the image of the latest monthly dyes!\n\n\n" +
                             "Note: This can be late so check the title of the image.");
 
-            Parser.parseFooter(embed, messageCreateEvent);
             Parser.parseColor(embed, messageCreateEvent);
             embed.setImage(Values.getDyeImage());
+
+            int hours = period.getHours();
+            int minutes = period.getMinutes();
+
+            embed.setFooter("Last check was " + hours + (hours == 1 ? " hour" : " hours") + " and " +
+                    minutes + (minutes == 1 ? " minute" : " minutes") + " ago.");
 
             messageCreateEvent.getChannel().sendMessage(embed);
         };